requirements.txt
apache-airflow
awswrangler
google-cloud-bigquery
requests
pandas
